pyproject.toml
[project]
name = "stockaivo"
version = "0.1.0"
description = "智能美股数据与分析平台后端"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.13",
    "uvicorn[standard]>=0.34.3",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    "redis>=6.2.0",
    "akshare>=1.17.6",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.1.0",
    "alembic>=1.16.2",
    "httpx>=0.28.1",
    "aiofiles>=24.1.0",
    "langgraph>=0.4.8",
    "langchain-core>=0.3.65",
    "google-generativeai>=0.8.5",
    "pandas-market-calendars>=5.1.1",
    "tenacity>=9.1.2",
    "apscheduler>=3.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

[project.scripts]
dev = "stockaivo.scripts.run:dev"
start = "stockaivo.scripts.run:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stockaivo"]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
